+++ .env
@@ -1,17 +1,11 @@
 ##
 ## Hasura
 ##
-ACCOUNT_EMAIL_WEBHOOK_URL=http://localhost:3000/api/webhooks/account
+ACCOUNT_EMAIL_WEBHOOK_URL=http://host.docker.internal:3000/api/webhooks/account
 
 FRONTEND_URL=http://localhost:3000
 GRAPHQL_ENDPOINT=http://localhost:8080/v1/graphql
 PORT=3000
-
-##
-## postgres
-##
-
-POSTGRES_PASSWORD=postgrespassword
 
 ##
 ## frontend secrets
@@ -39,9 +33,9 @@ NEXT_PUBLIC_ACTIVATION_TOKEN_EXPIRES=10080
 
 # Hasura
 HASURA_GRAPHQL_ADMIN_SECRET=admin1
-HASURA_GRAPHQL_DATABASE_URL=postgres://postgres:postgrespassword@localhost:5432/postgres
-HASURA_GRAPHQL_ENABLE_CONSOLE="true"
-HASURA_GRAPHQL_ENABLED_LOG_TYPES="startup, http-log, webhook-log, websocket-log, query-log"
+HASURA_GRAPHQL_DATABASE_URL=postgres://postgres:postgrespassword@postgres:5432/postgres
+HASURA_GRAPHQL_ENABLE_CONSOLE=true
+HASURA_GRAPHQL_ENABLED_LOG_TYPES=startup, http-log, webhook-log, websocket-log, query-log
 HASURA_GRAPHQL_JWT_SECRET={"type": "HS256", "key": "a_pretty_long_secret_key_that_should_be_at_least_32_char"}
 
 # webhook
