+++ targets/alert/.k8s/alert.cronjob.yaml
@@ -10,7 +10,7 @@ spec:
   concurrencyPolicy: Forbid
   successfulJobsHistoryLimit: 3
   failedJobsHistoryLimit: 3
-  schedule: "*/15 * * * *"
+  schedule: "0 2 * * *"
   jobTemplate:
     spec:
       backoffLimit: 0
